WORKING on flask project
+ CREATE flask application for back-end, the application can then be the back-end for any front-end application that
the user want to connect into.

+++ website/resources/item.py
@@ -3,6 +3,7 @@ from sqlalchemy.exc import SQLAlchemyError
 from flask import request
 from flask.views import MethodView
 from flask_smorest import abort, Blueprint
+from flask_jwt_extended import jwt_required, get_jwt
 from website.models import item
 from website.resources.schema import ItemSchema, ItemUpdateSchema
 from website.models import ItemModel
@@ -12,13 +13,19 @@ blp = Blueprint('Items', 'items', description='Operation on items')
 
 @blp.route('/item/<string:item_id>')
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self, item_id):
-        return ItemModel.query.all()
+        return ItemModel.query.filter_by(item_id)
     
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get('is_admin'):
+            abort(http_status_code=401, message='Admin priviledge is required...')
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
+        db.session.commit()
         return {'message': 'Item get removed'}, 200
     
     @blp.arguments(ItemUpdateSchema)
@@ -38,10 +45,12 @@ class Item(MethodView):
 
 @blp.route('/item')
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
-        return item.values()
+        return ItemModel.query.all()
     
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
